♻️  [Refactor]: Rename template files for consistency
template의 파일명 일관성을 위해 ComponentNameTemplate.tsx -> ComponentName.template.tsx로 변경
