+++ src/components/pages/house/house-regist/HouseRegister.tsx
@@ -11,8 +11,8 @@ import { SignUpProfileFormType } from '@/types/signUp.type';
 import Container from '@/components/atoms/Container';
 import Typography from '@/components/atoms/Typography';
 import Carousel from '@/components/organisms/Carousel';
-import HouseRegisterTemplate1 from '@/components/templates/house/house-regist/HouseRegisterTemplate1';
-import HouseRegisterTemplates2 from '@/components/templates/house/house-regist/HouseRegisterTemplates2';
+import HouseRegisterTemplate1 from '@/components/templates/house/house-regist/HouseRegister1.template';
+import HouseRegisterTemplates2 from '@/components/templates/house/house-regist/HouseRegister2.templates';
 import Button from '@/components/atoms/Button';
 import useModal from '@/hooks/useModal';
 import { InputRangeState } from '@/components/molecules/DualInputRange';
