+++ src/components/templates/house/house-detail/HouseImage.template.tsx
