+++ src/components/templates/house/house-detail/HouseInfoCard.tsx
@@ -4,7 +4,7 @@ import Divider from '@/components/atoms/Divider';
 import Icon from '@/components/atoms/Icon';
 import Typography from '@/components/atoms/Typography';
 import BadgeIcon from '@/components/molecules/BadgeIcon';
-import { HouseData } from '@/components/templates/house/house-detail/HouseDetailTemplate';
+import { HouseData } from '@/components/templates/house/house-detail/HouseDetail.template';
 import {
   floorInfo,
   genderInfo,
